Rename dao to model
