+++ libs/model/model_hash_provider_impl.cpp
@@ -96,5 +96,5 @@ namespace iroha {
       return concat_hash;
     }
 
-  }  // namespace dao
+  }  // namespace model
 }  // namespace iroha
